requirements.txt
tensorflow-gpu==2.1.0
opencv-python==4.0.0.21
tqdm==4.38.0
pandas==0.23.4
matplotlib==3.0.2
numpy==1.16.0
scikit-learn==0.20.2